Added information button
Added an information button to the twitter subpage. This button has an information "i" in it to let the user know that the button will give them more information. Upon pressing the button, the user will be sent to the twitter developer page where they will see how they can build a rule.